fix window into untagged enum
