+++ Add-VSCodeDir.ps1
@@ -1,1 +1,2 @@
+Write-Warning "Careful doing find and replace if you've ran this lol"
 code -a "$Env:APPDATA\Code"
