+++ src/storage_json/folder.rs
@@ -1,9 +1,10 @@
-use crate::backup_folder_id::BackupFolderId;
 use serde::Deserialize;
 use serde::Serialize;
 
+use crate::storage_json::uri::Uri;
+
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct EmptyWindow {
-    pub backup_folder: BackupFolderId,
+pub struct Folder {
+    pub folder_uri: Uri,
 }
