+++ src/storage_json/mod.rs
@@ -1,0 +1,24 @@
+pub mod machine_id;
+pub mod paths;
+pub mod storage_json;
+pub mod theme;
+pub mod color;
+pub mod windows_state;
+pub mod uri;
+pub mod window_splash;
+pub mod color_info;
+pub mod layout_info;
+pub mod side;
+pub mod backup_workspaces;
+pub mod workspace;
+pub mod workspace_id;
+pub mod folder;
+pub mod empty_window;
+pub mod backup_folder_id;
+pub mod profile_associations;
+pub mod profile_dunder;
+pub mod telemetry_sqm_id;
+pub mod telemetry_dev_device_id;
+pub mod workspace_identifier;
+pub mod window;
+pub mod ui_state;
