+++ src/storage_json/profile_associations.rs
@@ -2,9 +2,9 @@ use std::collections::HashMap;
 
 use serde::Deserialize;
 use serde::Serialize;
-use crate::backup_folder_id::BackupFolderId;
-use crate::profile_dunder::ProfileDunder;
-use crate::uri::Uri;
+use crate::storage_json::backup_folder_id::BackupFolderId;
+use crate::storage_json::profile_dunder::ProfileDunder;
+use crate::storage_json::uri::Uri;
 
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
