+++ src/storage_json/ui_state.rs
@@ -1,0 +1,13 @@
+use serde::Deserialize;
+use serde::Serialize;
+
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiState {
+    pub mode: u32,
+    pub x: u32,
+    pub y: u32,
+    pub width: u32,
+    pub height: u32,
+    pub zoom_level: Option<u8>,
+}
