+++ src/storage_json/workspace.rs
@@ -1,8 +1,8 @@
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::uri::Uri;
-use crate::workspace_id::WorkspaceId;
+use crate::storage_json::uri::Uri;
+use crate::storage_json::workspace_id::WorkspaceId;
 
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
