+++ vscodehelper-macros/Cargo.toml
@@ -16,4 +16,5 @@ quote = "1.0.40"
 syn = "2.0.100"
 
 [dev-dependencies]
+eyre = "0.6.12"
 serde = "1.0.219" # the crates that depend on this crate will also need to depend on serde I think lol
